+++ Runtime/UI/Manipulators/Pressable.cs
@@ -170,6 +170,9 @@ namespace Unity.AppUI.UI
             target.RegisterCallback<PointerUpEvent>(OnPointerUp);
             target.RegisterCallback<PointerCancelEvent>(OnPointerCancel);
             target.RegisterCallback<PointerCaptureOutEvent>(OnPointerCaptureOut);
+#if !UNITY_2023_1_OR_NEWER
+            target.RegisterCallback<MouseDownEvent>(OnMouseDown);
+#endif
             target.RegisterCallback<KeyDownEvent>(OnKeyDown);
             target.RegisterCallback<KeyUpEvent>(OnKeyUp);
         }
@@ -186,6 +189,9 @@ namespace Unity.AppUI.UI
             target.UnregisterCallback<PointerUpEvent>(OnPointerUp);
             target.UnregisterCallback<PointerCancelEvent>(OnPointerCancel);
             target.UnregisterCallback<PointerCaptureOutEvent>(OnPointerCaptureOut);
+#if !UNITY_2023_1_OR_NEWER
+            target.UnregisterCallback<MouseDownEvent>(OnMouseDown);
+#endif
             target.UnregisterCallback<KeyDownEvent>(OnKeyDown);
             target.UnregisterCallback<KeyUpEvent>(OnKeyUp);
         }
@@ -277,6 +283,16 @@ namespace Unity.AppUI.UI
             Activate(evt.pointerId);
             ProcessDownEvent(evt, evt.localPosition, evt.pointerId);
             evt.StopPropagation();
+        }
+        
+        void OnMouseDown(MouseDownEvent evt)
+        {
+            if (active)
+            {
+                if (!target.HasMouseCapture())
+                    target.CaptureMouse();
+                evt.StopPropagation();
+            }
         }
 
         void OnPointerMove(PointerMoveEvent evt)
@@ -376,7 +392,14 @@ namespace Unity.AppUI.UI
         void Activate(int pointerId)
         {
             if (!target.HasPointerCapture(pointerId))
+            {
                 target.CapturePointer(pointerId);
+#if !UNITY_2023_1_OR_NEWER
+                if (pointerId == PointerId.mousePointerId)
+                    target.CaptureMouse();
+#endif
+            }
+            
             ForceActivePseudoState();
             target.AddToClassList(Styles.activeUssClassName);
             m_PointerId = pointerId;