+++ Runtime/UI/Manipulators/Scrollable.cs
@@ -124,7 +124,14 @@ namespace Unity.AppUI.UI
             {
                 m_PointerId = evt.pointerId;
                 m_LastPos = evt.position;
-                target.CapturePointer(evt.pointerId);
+                if (!target.HasPointerCapture(evt.pointerId))
+                {
+                    target.CapturePointer(evt.pointerId);
+#if !UNITY_2023_1_OR_NEWER
+                    if (evt.pointerId == PointerId.mousePointerId)
+                        target.CaptureMouse();
+#endif
+                }
                 var pseudoStates = target.GetPseudoStates();
                 target.SetPseudoStates(pseudoStates | PseudoStates.Active);
             }